backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
openai
python-multipart
httpx
pillow
pytesseract
easyocr
markdown
playwright

